AMQP-119: Add cluster supportability for connector. Review SonarQ. Remove jdk6 and jdk7 supportability (not needed)

+++ mule-transport-amqp/src/test/java/org/mule/transport/amqp/internal/domain/AmqpMuleMessageFactoryTestCase.java
@@ -95,7 +95,7 @@ public class AmqpMuleMessageFactoryTestCase extends AbstractMuleContextTestCase
 
         final Envelope envelope = amqpMessage.getEnvelope();
         assertEquals(envelope.getDeliveryTag(),
-            muleMessage.getProperty(AmqpConnector.MESSAGE_PROPERTY_DELIVERY_TAG, PropertyScope.INBOUND));
+            (long) muleMessage.getProperty(AmqpConnector.MESSAGE_PROPERTY_DELIVERY_TAG, PropertyScope.INBOUND));
         assertEquals(envelope.isRedeliver(),
             muleMessage.getProperty(AmqpConnector.MESSAGE_PROPERTY_REDELIVER, PropertyScope.INBOUND));
         assertEquals(envelope.getExchange(),
